Remove extra delay for Dogecoin
This change removes the extra iteration that was causing the Dogecoin
cryptocurrency to show on the display for twice as long as the others.
This is because when we've reached the end of the entries in the cryptos
array, we reset the count to zero but do not update the display. This
change simplifies the incrementation logic and ensures we don't iterate
more than once per cryptocurrency.